Create kupyna_constants.h
Here's a basic structure for the kupyna_constants.h file which contains constants and lookup tables required for the Kupyna algorithm. Please note that the actual values for the constants and lookup tables are based on the official specification of the Kupyna hashing algorithm.

Please note that you'll need to fill in the actual values for the S-Box, round constants, and MDS matrix based on the official Kupyna specification, which can be found in the DSTU 7564:2014 standard. This is just a skeleton to help you structure your implementation.