Create kupyna_transform.cpp
The kupyna_transform.h and kupyna_transform.cpp files will contain the transformation logic of Kupyna, which is based on the T⊕ and T+ permutations borrowed from the Kalyna cipher. These functions form the core of the Kupyna hash function.

Please note that this is a simplified example and does not provide a full or optimized implementation. It is meant to give you a starting point on how you can structure the transformation functions based on the Kupyna specification. Additionally, some parts such as multiplication in the Galois field have been left out, and it's important to carefully implement and test these functions according to the official Kupyna specification which can be found in the DSTU 7564:2014 standard.