Create kalyna_constants.h
The kalyna_constants.h file contains constants and lookup tables required for the Kalyna cipher, which is used within the Kupyna hashing algorithm. This is a skeleton file and the actual values of the S-boxes and other constants should be filled in according to the official specification of the Kalyna cipher, which can be found in the DSTU 7624:2014 standard.

As with the Kupyna constants, it's important to carefully fill in the actual values for the S-boxes, round constants, and MDS matrix based on the official Kalyna specification in order to create a compliant and secure implementation.