Create main.cpp
Here is the main.cpp file, which will contain the main entry point of your program. This example will demonstrate how to use the Kupyna hashing algorithm to hash an input message. Please note that this is a high-level example, and assumes that the complete Kupyna implementation and all utility functions are in place.

In this example, the program takes an input message and passes it to the Kupyna hashing function. The resulting hash is then output in hexadecimal format.

Please note that this example assumes that a kupyna.h header file and corresponding implementation have been created, which define a kupyna_hash function. This function should take the input message, message length, output buffer, and desired hash size as parameters and perform the Kupyna hashing algorithm. This kupyna.h should tie together all the components, including constants, utility functions, and transformations, that have been discussed earlier. You would need to create and implement this file to have a working Kupyna hash function.