Create CMakeLists.txt
To compile the project with this CMakeLists.txt, you would typically run the following commands in the directory containing the CMakeLists.txt file:

sh
mkdir build
cd build
cmake ..
make

This will generate an executable named KupynaHash inside the build directory. Please ensure that all the necessary .cpp and .h files are in the same directory as the CMakeLists.txt file or adjust the paths in the CMakeLists.txt file accordingly.