Create kalyna_constantsV1.h
In the "A New Encryption Standard of Ukraine: The Kalyna Block Cipher" document, various constants such as the S-box and MDS matrix are defined. Here's how the kalyna_constants.h file can be refactored to include these constants:

The KALYNA_SBOX array should be filled with the values of the S-box as defined in the Kalyna specification. Similarly, the KALYNA_MDS_MATRIX and KALYNA_ROUND_CONSTANTS arrays should be filled with the respective values as defined in the specification.

As the document is quite extensive and detailed, it is important to ensure that all the constants are correctly transcribed from the specification. Also, keep in mind that implementing cryptographic algorithms is a delicate task and should be done with great care, ensuring that the implementation adheres to the specification.